requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
kafka-python
